made index a landing page and redirect from home when not logged in

+++ submitter/views.py
@@ -45,31 +45,36 @@ User = get_user_model()
 
 # send email with verification link
 def verify_email(request):
-    if request.method == "POST":
-        if request.user.email_is_verified != True:
-            current_site = get_current_site(request)
-            user = request.user
-            email = request.user.email
-            subject = "Verify Email"
-            message = render_to_string('submitter/verify_email_message.html', {
-                'request': request,
-                'user': user,
-                'domain': current_site.domain,
-                'uidb64':urlsafe_base64_encode(force_bytes(user.pk)),
-                'token':account_activation_token.make_token(user),
-            })
-            email = EmailMessage(
-                subject, message, to=[email]
-            )
-            email.content_subtype = 'html'
-            email.send()
-            return redirect('submitter:verify-email-done')
-        else:
-            return redirect('submitter:signup')
-    return render(request, 'submitter/verify_email.html')
+    if request.user.is_authenticated:
+        if request.method == "POST":
+            if request.user.email_is_verified != True:
+                current_site = get_current_site(request)
+                user = request.user
+                email = request.user.email
+                subject = "Verify Email"
+                message = render_to_string('submitter/verify_email_message.html', {
+                    'request': request,
+                    'user': user,
+                    'domain': current_site.domain,
+                    'uidb64':urlsafe_base64_encode(force_bytes(user.pk)),
+                    'token':account_activation_token.make_token(user),
+                })
+                email = EmailMessage(
+                    subject, message, to=[email]
+                )
+                email.content_subtype = 'html'
+                email.send()
+                return redirect('submitter:verify-email-done')
+            else:
+                return redirect('submitter:signup')
+        return render(request, 'submitter/verify_email.html')
+    else:
+        return redirect('submitter:login')
 
 def index(request):
-    return redirect('submitter:register')
+    if request.user.is_authenticated:
+        return redirect("submitter:home")
+    return render(request, 'submitter/landing.html')
 
 def submission(request, listing_id):
     listing = Listing.objects.get(pk = listing_id)
@@ -110,7 +115,10 @@ def submission(request, listing_id):
     return render(request, "submitter/submission.html", context)
 
 def results(request, listing_id):
+
     listing = Listing.objects.get(id = listing_id)
+    if not listing:
+        return redirect("submitter:home")
     if not request.user.id == listing.creator.id:
         return redirect("submitter:home")
 
@@ -167,18 +175,24 @@ def results(request, listing_id):
 
 def close_listing(request, listing_id):
     listing = Listing.objects.get(pk = listing_id)
+    if not request.user.id == listing.creator.id:
+        return redirect("submitter:home")
     listing.is_closed = True
     listing.save()
     return redirect('submitter:results', listing_id)
 
 def reopen_listing(request, listing_id):
     listing = Listing.objects.get(pk = listing_id)
+    if not request.user.id == listing.creator.id:
+        return redirect("submitter:home")
     listing.is_closed = False
     listing.save()
     return redirect('submitter:results', listing_id)
 
 def delete_listing(request, listing_id):
     listing = Listing.objects.get(pk = listing_id)
+    if not request.user.id == listing.creator.id:
+        return redirect("submitter:home")
     listing.delete()
     return redirect('submitter:home')
 
@@ -327,7 +341,7 @@ def registerPage(request):
                 if "submit" in request.session:
                     submit_from_redirect(request, user)
                 login(request, user)
-                return redirect('submitter:verify-email')
+                return verify_email(request)
             else:
                 context = {'form': form}
                 return render(request, "submitter/register.html", context)
@@ -376,7 +390,7 @@ def homePage(request):
     elif request.user.is_authenticated:
         return redirect('submitter:verify-email')
     else:
-        return redirect('submitter:login')
+        return redirect('submitter:index')
 
 
 def logout_view(request):
